+++ examples/sample_benchmark.py
@@ -27,9 +27,8 @@ for n in [8, 10, 12, 14, 16]:
         print("n: ", n, " nlayers: ", nlayers)
         c = construct_circuit(n, nlayers)
         time0 = time.time()
-        s = c.state()
+        s = c.sample(allow_state=True)
         time1 = time.time()
-        smp = bin(np.random.choice(range(2**n), p=np.abs(K.numpy(s)) ** 2))
         # print(smp)
         print("state sampling time: ", time1 - time0)
         time0 = time.time()
@@ -37,6 +36,10 @@ for n in [8, 10, 12, 14, 16]:
         # print(smp)
         time1 = time.time()
         print("nonjit tensor sampling time: ", time1 - time0)
+        time0 = time.time()
+        s = c.sample(allow_state=True, batch=10)
+        time1 = time.time()
+        print("batch state sampling time: ", (time1 - time0) / 10)
 
         @K.jit
         def f(key):
@@ -48,11 +51,10 @@ for n in [8, 10, 12, 14, 16]:
         time0 = time.time()
         smp = f(key1)
         time1 = time.time()
-        for _ in range(5):
+        for _ in range(10):
             key1, key2 = K.random_split(key2)
             smp = f(key1)
             # print(smp)
         time2 = time.time()
-
-        print("jittable tensor sampling staging time: ", time1 - time0)
-        print("jittable tensor sampling running time: ", (time2 - time1) / 5)
+        print("jittable tensor sampling staginging time: ", time1 - time0)
+        print("jittable tensor sampling running time: ", (time2 - time1) / 10)